+++ benchmark/engine/CMakeLists.txt
@@ -12,12 +12,14 @@ FetchContent_MakeAvailable(nlohmann_json)
 
 set(engine_benchmark_srcs
   ${CMAKE_CURRENT_SOURCE_DIR}/scenario_dispatcher.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/common/benchmark_types.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/common/utils.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/utils.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/scenario_base.h
   ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/scenario_base.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/decode_baseline.h
   ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/decode_baseline.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/long_prefill.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/scenarios/long_prefill.cpp
 )
 
 add_executable(engine_benchmark ${engine_benchmark_srcs})
