+++ benchmark/engine/config.json
@@ -9,5 +9,16 @@
     "execution_provider_library": "build/Linux/RelWithDebInfo/libonnxruntime_providers_cuda.so",
     "generation_tokens": 128,
     "measured_runs": 2
+  },
+  {
+    "scenario": "long_prefill",
+    "concurrency": 1,
+    "prompt_length_k": 32,
+    "synthetic": true,
+    "model_path": "C:/models/qwen2.5-0.5b-instruct",
+    "execution_provider": "cuda",
+    "execution_provider_library": "build/Linux/RelWithDebInfo/libonnxruntime_providers_cuda.so",
+    "generation_tokens": 8,
+    "measured_runs": 2
   }
 ]
